chore: add no-op middleware outside of Node.js
