+++ packages/middleware-recursion-detection/src/configuration.ts
@@ -1,3 +1,12 @@
-export interface PreviouslyResolved {
-  runtime: string;
-}
+import { AbsoluteLocation, BuildHandlerOptions } from "@smithy/types";
+
+/**
+ * @internal
+ */
+export const recursionDetectionMiddlewareOptions: BuildHandlerOptions & AbsoluteLocation = {
+  step: "build",
+  tags: ["RECURSION_DETECTION"],
+  name: "recursionDetectionMiddleware",
+  override: true,
+  priority: "low",
+};
