+++ packages/middleware-recursion-detection/src/recursionDetectionMiddleware.no-op.ts
@@ -1,0 +1,17 @@
+import {
+  BuildHandler,
+  BuildHandlerArguments,
+  BuildHandlerOutput,
+  BuildMiddleware,
+  MetadataBearer,
+} from "@smithy/types";
+
+/**
+ * No-op middleware for runtimes outside of Node.js
+ * @internal
+ */
+export const recursionDetectionMiddleware =
+  (): BuildMiddleware<any, any> =>
+  <Output extends MetadataBearer>(next: BuildHandler<any, Output>): BuildHandler<any, Output> =>
+  async (args: BuildHandlerArguments<any>): Promise<BuildHandlerOutput<Output>> =>
+    next(args);
